TenGigEth.xdc: added false path for ethRefClk -> txClockGt
These clocks are found to be unrelated by the clock interaction
report. They have the same source clock and period but txClockGt
is divided down from a PLL and thus might have a different phase.

I verified that all CDC paths are synchronized (except for
a few minor status bits that go into the AXI registers and
for which I filed PR #392).